Get, post, put and delete seems to be working. Int instead of double on properties.

+++ src/main/java/dat3/AXP/dto/EquipmentDto.java
@@ -12,8 +12,8 @@ public class EquipmentDto {
     private Integer id;
     private String name;
     private boolean status;
-    private double totalUnits;
-    private double defectiveUnits;
+    private int totalUnits;
+    private int defectiveUnits;
     private int activityId;
